requirements.txt
pandas
requests
faiss-cpu
numpy
python-dotenv
flask